pyproject.toml
[project]
name = "tensorrt-llm-model-user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "locust>=2.37.1",
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "result>=0.17.0",
    "tritonclient>=2.46.0",
]

[dependency-groups]
ci = [
    "mypy>=1.15.0",
    "ruff>=0.11.9",
    "types-requests>=2.32.0.20250328",
]
